move api hooks to its own directory
