+++ frontend/mulighetsrommet-veileder-flate/src/api/queries/useTiltaksgjennomforing.ts
@@ -1,5 +1,5 @@
 import { useSanity } from '../useSanity';
-import { Tiltaksgjennomforing } from '../../api/models';
+import { Tiltaksgjennomforing } from '../models';
 
 export default function useTiltaksgjennomforing() {
   return useSanity<Tiltaksgjennomforing[]>(`*[_type == "tiltaksgjennomforing"]{
