+++ frontend/mulighetsrommet-veileder-flate/src/api/queries/useTiltaksgjennomforingById.ts
@@ -1,5 +1,5 @@
 import { useSanity } from '../useSanity';
-import { Tiltaksgjennomforing } from '../../api/models';
+import { Tiltaksgjennomforing } from '../models';
 
 export default function useTiltaksgjennomforingById(id: number) {
   return useSanity<Tiltaksgjennomforing[]>(`*[_type == "tiltaksgjennomforing" && tiltaksnummer == ${id}]{
