+++ frontend/mulighetsrommet-veileder-flate/src/api/useSanity.ts
@@ -1,6 +1,6 @@
 import { useQuery } from 'react-query';
 import { QueryKeys } from '../core/api/QueryKeys';
-import { client } from '../api/sanityClient';
+import { client } from './sanityClient';
 
 export function useSanity<T>(groqQuery: string) {
   return useQuery([QueryKeys.SanityQuery, groqQuery], () => client.query<T>(groqQuery));
